NET-4527 ShimLayer Generator: Remove ToConversionSnippet
GitOrigin-RevId: 388660abff5bc0d3775970b64a92180c56372716

+++ analyzers/tests/SonarAnalyzer.ShimLayer.Generator.Test/Snapshots/Snap#ILabelSymbol.g.cs.verified.cs
@@ -26,6 +26,6 @@ public static class ILabelSymbolShimExtensions
 
     extension(ILabelSymbol wrappedInstance)
     {
-        public int MetadataToken => (int)MetadataTokenAccessor(wrappedInstance);
+        public int MetadataToken => MetadataTokenAccessor(wrappedInstance);
     }
 }
