+++ analyzers/tests/SonarAnalyzer.ShimLayer.Generator.Test/Snapshots/Snap#IPreprocessingSymbol.g.cs.verified.cs
@@ -26,6 +26,6 @@ public static class IPreprocessingSymbolShimExtensions
 
     extension(IPreprocessingSymbol wrappedInstance)
     {
-        public int MetadataToken => (int)MetadataTokenAccessor(wrappedInstance);
+        public int MetadataToken => MetadataTokenAccessor(wrappedInstance);
     }
 }
